server/requirements.txt
Flask
Flask-CORS
requests
fuzzywuzzy
dnspython
bs4